Move image functions to subfolders
