+++ functions/src/image/fetch-image/query-constants.ts
